creation du header + makefile

+++ ft_memchr.c
@@ -6,13 +6,15 @@
 /*   By: fzaazaa <fzaazaa@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/10 10:11:00 by fzaazaa           #+#    #+#             */
-/*   Updated: 2024/11/10 10:55:46 by fzaazaa          ###   ########.fr       */
+/*   Updated: 2024/11/23 16:05:44 by fzaazaa          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memchr(const void *s, int c, int n)
+#include"libft.h"
+
+void	*ft_memchr(const void *s, int c, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
