+++ ft_strncmp.c
@@ -6,13 +6,15 @@
 /*   By: fzaazaa <fzaazaa@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/02 13:10:39 by fzaazaa           #+#    #+#             */
-/*   Updated: 2024/11/10 10:55:58 by fzaazaa          ###   ########.fr       */
+/*   Updated: 2024/11/23 15:51:14 by fzaazaa          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strncmp(char *s1, char *s2, unsigned int n)
+#include"libft.h"
+
+int	ft_strncmp(char *s1, char *s2, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (s1[i] && i < n)
